delete store folder and move 'store.ts' file to src folder
