+++ frontend/__tests__/src/App.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import App from '@src/App'
 import { Provider } from 'react-redux'
-import { store } from '@src/store/store'
+import { store } from '@src/store'
 
 describe('render app', () => {
   const setup = () => {
