+++ frontend/src/features/board/boardSlice.tsx
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit'
-import type { RootState } from '@src/store/store'
+import type { RootState } from '@src/store'
 
 export interface boardState {
   isBoardVerified: boolean
