+++ frontend/src/features/stepper/StepperSlice.tsx
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit'
-import type { RootState } from '@src/store/store'
+import type { RootState } from '@src/store'
 import { ZERO } from '@src/constants'
 
 export interface StepState {
