+++ frontend/src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 import { Provider } from 'react-redux'
-import { store } from './store/store'
+import { store } from './store'
 import { ThemeProvider } from '@mui/material'
 import { theme } from '@src/theme'
