+++ frontend/src/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit'
-import stepperReducer from '../features/stepper/StepperSlice'
-import configReducer from '../features/config/configSlice'
-import boardReducer from '../features/board/boardSlice'
+import stepperReducer from './features/stepper/StepperSlice'
+import configReducer from './features/config/configSlice'
+import boardReducer from './features/board/boardSlice'
 export const store = configureStore({
   reducer: {
     stepper: stepperReducer,
